Generate an HTML Report File

This function creates and writes an HTML report file based on specified study and structure details.

write_html_file(
  path,
  structure = "USMR",
  path_logo = NULL,
  confidential = FALSE,
  report_type = "Data review report",
  study_id = "CHUBXYYYY/NN",
  study_name = "The Study Name",
  study_abbreviation = "TSN",
  investigator = "Investigator name",
  methodologist = "Jean Dupont",
  biostatistician = "George Frais",
  datamanager = "Peter Parker",
  methodologist_mail = NULL,
  biostatistician_mail = NULL,
  datamanager_mail = NULL,
  language = "fr"
)

Arguments

path

Character string specifying the file path where the HTML file will be saved.

structure

Character string indicating the organizational structure, either "USMR" or "EUCLID". Default is "USMR".

path_logo

Character string specifying the path to the logo image. If NULL, a default logo is used.

confidential

Logical value indicating whether the report should be marked as confidential. Default is FALSE.

report_type

Character string specifying the type of report. Default is "Data review report".

study_id

Character string representing the study identifier. Default is "CHUBXYYYY/NN".

study_name

Character string specifying the name of the study. Default is "The Study Name".

study_abbreviation

Character string providing the abbreviation of the study. Default is "TSN".

investigator

Character string representing the investigator's name. Default is "Investigator name".

methodologist

Character string specifying the methodologist's name. Default is "Jean Dupont".

biostatistician

Character string specifying the biostatistician's name. Default is "George Frais".

datamanager

Character string specifying the data manager's name. Default is "Peter Parker".

methodologist_mail

Character string specifying the methodologist's email. If NULL, it is generated automatically.

biostatistician_mail

Character string specifying the biostatistician's email. If NULL, it is generated automatically.

datamanager_mail

Character string specifying the data manager's email. If NULL, it is generated automatically.

language

Character string indicating the language of the report, either "fr" (French) or "en" (English). Default is "fr".

Value

None. The function writes an HTML report to the specified file path.